Refactor: Convert BedService back to BedManager as it is not a service.
